Refactor structure apps/readmanga_parser -> apps/parse/readmanga
