+++ apps/parse/readmanga/readmanga/descr_utils.py
