+++ apps/parse/readmanga/readmanga/spiders/consts.py
