+++ apps/parse/readmanga/scrapy.cfg
@@ -1,0 +1,11 @@
+# Automatically created by: scrapy startproject
+#
+# For more information about the [deploy] section see:
+# https://scrapyd.readthedocs.io/en/latest/deploy.html
+
+[settings]
+default = apps.parse.readmanga.readmanga.settings
+
+[deploy]
+#url = http://localhost:6800/
+project = readmanga
