+++ apps/parse/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from apps.parse.views import SearchAPIView
+
+urlpatterns = [
+    path("search/<str:title>", SearchAPIView.as_view()),
+]
