--- apps/readmanga_parser/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class ReadmangaParserConfig(AppConfig):
-    name = "apps.readmanga_parser"
-    verbose_name = "Readmanga"
