+++ manga_reader/settings.py
@@ -36,7 +36,7 @@ INSTALLED_APPS = [
     "corsheaders",
     "apps.api_docs",
     "apps.login.apps.LoginConfig",
-    "apps.readmanga_parser.apps.ReadmangaParserConfig",
+    "apps.parse",
     "django_extensions",
     "django.contrib.postgres",
 ]
@@ -52,7 +52,7 @@ JAZZMIN_SETTINGS = {
     "site_logo": "favicon.ico",
     "welcome_sign": "Admin panel",
     "copyright": '<a target="_blank" href="https://github.com/sora-reader">Sora</a>',
-    "search_model": "readmanga_parser.manga",
+    "search_model": "parse.manga",
     "user_avatar": None,
     ############
     # Top Menu #
@@ -74,18 +74,18 @@ JAZZMIN_SETTINGS = {
     "hide_apps": ["token_blacklist"],
     "hide_models": ["auth.group"],
     "icons": {
-        "readmanga_parser.manga": "fas fa-book-open",
-        "readmanga_parser.author": "fas fa-user-edit",
-        "readmanga_parser.genre": "fas fa-theater-masks",
-        "readmanga_parser.translator": "fas fa-language",
+        "parse.manga": "fas fa-book-open",
+        "parse.author": "fas fa-user-edit",
+        "parse.genre": "fas fa-theater-masks",
+        "parse.translator": "fas fa-language",
         "auth.user": "fas fa-user",
         "auth.group": "fas fa-users",
     },
     "order_with_respect_to": [
-        "readmanga_parser",
-        "readmanga_parser.manga",
-        "readmanga_parser.author",
-        "readmanga_parser.genre",
+        "parse",
+        "parse.manga",
+        "parse.author",
+        "parse.genre",
         "auth",
         "login",
         "authtoken",
