+++ manga_reader/urls.py
@@ -4,7 +4,7 @@ from django.urls import path
 
 apipatterns = [
     path("docs/", include("apps.api_docs.urls")),
-    path("readmanga/", include("apps.readmanga_parser.urls")),
+    path("readmanga/", include("apps.parse.urls")),
     path("auth/", include("apps.login.urls")),
 ]
